+++ src/query/range_query/range_query.rs
@@ -1,21 +1,17 @@
 use std::io;
-use std::net::Ipv6Addr;
-use std::ops::{Bound, Range};
+use std::ops::Bound;
 
-use columnar::MonotonicallyMappableToU128;
-use common::{BinarySerializable, BitSet};
+use common::BitSet;
 
 use super::map_bound;
 use super::range_query_u64_fastfield::FastFieldRangeWeight;
-use crate::error::TantivyError;
 use crate::index::SegmentReader;
 use crate::query::explanation::does_not_match;
-use crate::query::range_query::range_query_ip_fastfield::IPFastFieldRangeWeight;
-use crate::query::range_query::{is_type_valid_for_fastfield_range_query, map_bound_res};
+use crate::query::range_query::is_type_valid_for_fastfield_range_query;
 use crate::query::{BitSetDocSet, ConstScorer, EnableScoring, Explanation, Query, Scorer, Weight};
 use crate::schema::{Field, IndexRecordOption, Term, Type};
 use crate::termdict::{TermDictionary, TermStreamer};
-use crate::{DateTime, DocId, Score};
+use crate::{DocId, Score};
 
 /// `RangeQuery` matches all documents that have at least one term within a defined range.
 ///
@@ -40,8 +36,10 @@ use crate::{DateTime, DocId, Score};
 /// ```rust
 /// use tantivy::collector::Count;
 /// use tantivy::query::RangeQuery;
+/// use tantivy::Term;
 /// use tantivy::schema::{Schema, INDEXED};
 /// use tantivy::{doc, Index, IndexWriter};
+/// use std::ops::Bound;
 /// # fn test() -> tantivy::Result<()> {
 /// let mut schema_builder = Schema::builder();
 /// let year_field = schema_builder.add_u64_field("year", INDEXED);
@@ -59,7 +57,10 @@ use crate::{DateTime, DocId, Score};
 ///
 /// let reader = index.reader()?;
 /// let searcher = reader.searcher();
-/// let docs_in_the_sixties = RangeQuery::new_u64("year".to_string(), 1960..1970);
+/// let docs_in_the_sixties = RangeQuery::new(
+///     Bound::Included(Term::from_field_u64(year_field, 1960)),
+///     Bound::Excluded(Term::from_field_u64(year_field, 1970)),
+/// );
 /// let num_60s_books = searcher.search(&docs_in_the_sixties, &Count)?;
 /// assert_eq!(num_60s_books, 2285);
 /// Ok(())
@@ -68,11 +69,17 @@ use crate::{DateTime, DocId, Score};
 /// ```
 #[derive(Clone, Debug)]
 pub struct RangeQuery {
-    field: String,
-    value_type: Type,
-    lower_bound: Bound<Vec<u8>>,
-    upper_bound: Bound<Vec<u8>>,
+    lower_bound: Bound<Term>,
+    upper_bound: Bound<Term>,
     limit: Option<u64>,
+}
+
+/// Returns the inner value of a `Bound`
+pub(crate) fn inner_bound(val: &Bound<Term>) -> Option<&Term> {
+    match val {
+        Bound::Included(term) | Bound::Excluded(term) => Some(term),
+        Bound::Unbounded => None,
+    }
 }
 
 impl RangeQuery {
@@ -80,234 +87,28 @@ impl RangeQuery {
     ///
     /// If the value type is not correct, something may go terribly wrong when
     /// the `Weight` object is created.
-    pub fn new_term_bounds(
-        field: String,
-        value_type: Type,
-        lower_bound: &Bound<Term>,
-        upper_bound: &Bound<Term>,
-    ) -> RangeQuery {
-        let verify_and_unwrap_term = |val: &Term| val.serialized_value_bytes().to_owned();
+    pub fn new(lower_bound: Bound<Term>, upper_bound: Bound<Term>) -> RangeQuery {
         RangeQuery {
-            field,
-            value_type,
-            lower_bound: map_bound(lower_bound, verify_and_unwrap_term),
-            upper_bound: map_bound(upper_bound, verify_and_unwrap_term),
+            lower_bound,
+            upper_bound,
             limit: None,
         }
-    }
-
-    /// Creates a new `RangeQuery` over a `i64` field.
-    ///
-    /// If the field is not of the type `i64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_i64(field: String, range: Range<i64>) -> RangeQuery {
-        RangeQuery::new_i64_bounds(
-            field,
-            Bound::Included(range.start),
-            Bound::Excluded(range.end),
-        )
-    }
-
-    /// Create a new `RangeQuery` over a `i64` field.
-    ///
-    /// The two `Bound` arguments make it possible to create more complex
-    /// ranges than semi-inclusive range.
-    ///
-    /// If the field is not of the type `i64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_i64_bounds(
-        field: String,
-        lower_bound: Bound<i64>,
-        upper_bound: Bound<i64>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &i64| {
-            Term::from_field_i64(Field::from_field_id(0), *val)
-                .serialized_value_bytes()
-                .to_owned()
-        };
-        RangeQuery {
-            field,
-            value_type: Type::I64,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Creates a new `RangeQuery` over a `f64` field.
-    ///
-    /// If the field is not of the type `f64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_f64(field: String, range: Range<f64>) -> RangeQuery {
-        RangeQuery::new_f64_bounds(
-            field,
-            Bound::Included(range.start),
-            Bound::Excluded(range.end),
-        )
-    }
-
-    /// Create a new `RangeQuery` over a `f64` field.
-    ///
-    /// The two `Bound` arguments make it possible to create more complex
-    /// ranges than semi-inclusive range.
-    ///
-    /// If the field is not of the type `f64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_f64_bounds(
-        field: String,
-        lower_bound: Bound<f64>,
-        upper_bound: Bound<f64>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &f64| {
-            Term::from_field_f64(Field::from_field_id(0), *val)
-                .serialized_value_bytes()
-                .to_owned()
-        };
-        RangeQuery {
-            field,
-            value_type: Type::F64,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Create a new `RangeQuery` over a `u64` field.
-    ///
-    /// The two `Bound` arguments make it possible to create more complex
-    /// ranges than semi-inclusive range.
-    ///
-    /// If the field is not of the type `u64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_u64_bounds(
-        field: String,
-        lower_bound: Bound<u64>,
-        upper_bound: Bound<u64>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &u64| {
-            Term::from_field_u64(Field::from_field_id(0), *val)
-                .serialized_value_bytes()
-                .to_owned()
-        };
-        RangeQuery {
-            field,
-            value_type: Type::U64,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Create a new `RangeQuery` over a `ip` field.
-    ///
-    /// If the field is not of the type `ip`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_ip_bounds(
-        field: String,
-        lower_bound: Bound<Ipv6Addr>,
-        upper_bound: Bound<Ipv6Addr>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &Ipv6Addr| {
-            Term::from_field_ip_addr(Field::from_field_id(0), *val)
-                .serialized_value_bytes()
-                .to_owned()
-        };
-        RangeQuery {
-            field,
-            value_type: Type::IpAddr,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Create a new `RangeQuery` over a `u64` field.
-    ///
-    /// If the field is not of the type `u64`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_u64(field: String, range: Range<u64>) -> RangeQuery {
-        RangeQuery::new_u64_bounds(
-            field,
-            Bound::Included(range.start),
-            Bound::Excluded(range.end),
-        )
-    }
-
-    /// Create a new `RangeQuery` over a `date` field.
-    ///
-    /// The two `Bound` arguments make it possible to create more complex
-    /// ranges than semi-inclusive range.
-    ///
-    /// If the field is not of the type `date`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_date_bounds(
-        field: String,
-        lower_bound: Bound<DateTime>,
-        upper_bound: Bound<DateTime>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &DateTime| {
-            Term::from_field_date(Field::from_field_id(0), *val)
-                .serialized_value_bytes()
-                .to_owned()
-        };
-        RangeQuery {
-            field,
-            value_type: Type::Date,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Create a new `RangeQuery` over a `date` field.
-    ///
-    /// If the field is not of the type `date`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_date(field: String, range: Range<DateTime>) -> RangeQuery {
-        RangeQuery::new_date_bounds(
-            field,
-            Bound::Included(range.start),
-            Bound::Excluded(range.end),
-        )
-    }
-
-    /// Create a new `RangeQuery` over a `Str` field.
-    ///
-    /// The two `Bound` arguments make it possible to create more complex
-    /// ranges than semi-inclusive range.
-    ///
-    /// If the field is not of the type `Str`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_str_bounds(
-        field: String,
-        lower_bound: Bound<&str>,
-        upper_bound: Bound<&str>,
-    ) -> RangeQuery {
-        let make_term_val = |val: &&str| val.as_bytes().to_vec();
-        RangeQuery {
-            field,
-            value_type: Type::Str,
-            lower_bound: map_bound(&lower_bound, make_term_val),
-            upper_bound: map_bound(&upper_bound, make_term_val),
-            limit: None,
-        }
-    }
-
-    /// Create a new `RangeQuery` over a `Str` field.
-    ///
-    /// If the field is not of the type `Str`, tantivy
-    /// will panic when the `Weight` object is created.
-    pub fn new_str(field: String, range: Range<&str>) -> RangeQuery {
-        RangeQuery::new_str_bounds(
-            field,
-            Bound::Included(range.start),
-            Bound::Excluded(range.end),
-        )
     }
 
     /// Field to search over
-    pub fn field(&self) -> &str {
-        &self.field
+    pub fn field(&self) -> Field {
+        self.get_term().field()
+    }
+
+    /// The value type of the field
+    pub fn value_type(&self) -> Type {
+        self.get_term().typ()
+    }
+
+    pub(crate) fn get_term(&self) -> &Term {
+        inner_bound(&self.lower_bound)
+            .or(inner_bound(&self.upper_bound))
+            .expect("At least one bound must be set")
     }
 
     /// Limit the number of term the `RangeQuery` will go through.
@@ -319,70 +120,23 @@ impl RangeQuery {
     }
 }
 
-/// Returns true if the type maps to a u64 fast field
-pub(crate) fn maps_to_u64_fastfield(typ: Type) -> bool {
-    match typ {
-        Type::U64 | Type::I64 | Type::F64 | Type::Bool | Type::Date => true,
-        Type::IpAddr => false,
-        Type::Str | Type::Facet | Type::Bytes | Type::Json => false,
-    }
-}
-
 impl Query for RangeQuery {
     fn weight(&self, enable_scoring: EnableScoring<'_>) -> crate::Result<Box<dyn Weight>> {
         let schema = enable_scoring.schema();
-        let field_type = schema
-            .get_field_entry(schema.get_field(&self.field)?)
-            .field_type();
-        let value_type = field_type.value_type();
-        if value_type != self.value_type {
-            let err_msg = format!(
-                "Create a range query of the type {:?}, when the field given was of type \
-                 {value_type:?}",
-                self.value_type
-            );
-            return Err(TantivyError::SchemaError(err_msg));
-        }
+        let field_type = schema.get_field_entry(self.field()).field_type();
 
-        if field_type.is_fast() && is_type_valid_for_fastfield_range_query(self.value_type) {
-            if field_type.is_ip_addr() {
-                let parse_ip_from_bytes = |data: &Vec<u8>| {
-                    let ip_u128_bytes: [u8; 16] = data.as_slice().try_into().map_err(|_| {
-                        crate::TantivyError::InvalidArgument(
-                            "Expected 8 bytes for ip address".to_string(),
-                        )
-                    })?;
-                    let ip_u128 = u128::from_be_bytes(ip_u128_bytes);
-                    crate::Result::<Ipv6Addr>::Ok(Ipv6Addr::from_u128(ip_u128))
-                };
-                let lower_bound = map_bound_res(&self.lower_bound, parse_ip_from_bytes)?;
-                let upper_bound = map_bound_res(&self.upper_bound, parse_ip_from_bytes)?;
-                Ok(Box::new(IPFastFieldRangeWeight::new(
-                    self.field.to_string(),
-                    lower_bound,
-                    upper_bound,
-                )))
-            } else {
-                // We run the range query on u64 value space for performance reasons and simpicity
-                // assert the type maps to u64
-                assert!(maps_to_u64_fastfield(self.value_type));
-                let parse_from_bytes = |data: &Vec<u8>| {
-                    u64::from_be(BinarySerializable::deserialize(&mut &data[..]).unwrap())
-                };
-
-                let lower_bound = map_bound(&self.lower_bound, parse_from_bytes);
-                let upper_bound = map_bound(&self.upper_bound, parse_from_bytes);
-                Ok(Box::new(FastFieldRangeWeight::new_u64_lenient(
-                    self.field.to_string(),
-                    lower_bound,
-                    upper_bound,
-                )))
-            }
+        if field_type.is_fast() && is_type_valid_for_fastfield_range_query(self.value_type()) {
+            Ok(Box::new(FastFieldRangeWeight::new(
+                self.field(),
+                self.lower_bound.clone(),
+                self.upper_bound.clone(),
+            )))
         } else {
+            let verify_and_unwrap_term = |val: &Term| val.serialized_value_bytes().to_owned();
             Ok(Box::new(RangeWeight {
-                field: self.field.to_string(),
-                lower_bound: self.lower_bound.clone(),
-                upper_bound: self.upper_bound.clone(),
+                field: self.field(),
+                lower_bound: map_bound(&self.lower_bound, verify_and_unwrap_term),
+                upper_bound: map_bound(&self.upper_bound, verify_and_unwrap_term),
                 limit: self.limit,
             }))
         }
@@ -390,7 +144,7 @@ impl Query for RangeQuery {
 }
 
 pub struct RangeWeight {
-    field: String,
+    field: Field,
     lower_bound: Bound<Vec<u8>>,
     upper_bound: Bound<Vec<u8>>,
     limit: Option<u64>,
@@ -423,7 +177,7 @@ impl Weight for RangeWeight {
         let max_doc = reader.max_doc();
         let mut doc_bitset = BitSet::with_max_value(max_doc);
 
-        let inverted_index = reader.inverted_index(reader.schema().get_field(&self.field)?)?;
+        let inverted_index = reader.inverted_index(self.field)?;
         let term_dict = inverted_index.terms();
         let mut term_range = self.term_range(term_dict)?;
         let mut processed_count = 0;
@@ -477,7 +231,7 @@ mod tests {
     use crate::schema::{
         Field, IntoIpv6Addr, Schema, TantivyDocument, FAST, INDEXED, STORED, TEXT,
     };
-    use crate::{Index, IndexWriter};
+    use crate::{Index, IndexWriter, Term};
 
     #[test]
     fn test_range_query_simple() -> crate::Result<()> {
@@ -499,7 +253,10 @@ mod tests {
         let reader = index.reader()?;
         let searcher = reader.searcher();
 
-        let docs_in_the_sixties = RangeQuery::new_u64("year".to_string(), 1960u64..1970u64);
+        let docs_in_the_sixties = RangeQuery::new(
+            Bound::Included(Term::from_field_u64(year_field, 1960)),
+            Bound::Excluded(Term::from_field_u64(year_field, 1970)),
+        );
 
         // ... or `1960..=1969` if inclusive range is enabled.
         let count = searcher.search(&docs_in_the_sixties, &Count)?;
@@ -530,7 +287,10 @@ mod tests {
         let reader = index.reader()?;
         let searcher = reader.searcher();
 
-        let mut docs_in_the_sixties = RangeQuery::new_u64("year".to_string(), 1960u64..1970u64);
+        let mut docs_in_the_sixties = RangeQuery::new(
+            Bound::Included(Term::from_field_u64(year_field, 1960)),
+            Bound::Excluded(Term::from_field_u64(year_field, 1970)),
+        );
         docs_in_the_sixties.limit(5);
 
         // due to the limit and no docs in 1963, it's really only 1960..=1965
@@ -575,29 +335,29 @@ mod tests {
             |range_query: RangeQuery| searcher.search(&range_query, &Count).unwrap();
 
         assert_eq!(
-            count_multiples(RangeQuery::new_i64("intfield".to_string(), 10..11)),
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_i64(int_field, 10)),
+                Bound::Excluded(Term::from_field_i64(int_field, 11)),
+            )),
             9
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_i64_bounds(
-                "intfield".to_string(),
-                Bound::Included(10),
-                Bound::Included(11)
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_i64(int_field, 10)),
+                Bound::Included(Term::from_field_i64(int_field, 11)),
             )),
             18
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_i64_bounds(
-                "intfield".to_string(),
-                Bound::Excluded(9),
-                Bound::Included(10)
+            count_multiples(RangeQuery::new(
+                Bound::Excluded(Term::from_field_i64(int_field, 9)),
+                Bound::Included(Term::from_field_i64(int_field, 10)),
             )),
             9
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_i64_bounds(
-                "intfield".to_string(),
-                Bound::Included(9),
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_i64(int_field, 9)),
                 Bound::Unbounded
             )),
             91
@@ -646,29 +406,29 @@ mod tests {
             |range_query: RangeQuery| searcher.search(&range_query, &Count).unwrap();
 
         assert_eq!(
-            count_multiples(RangeQuery::new_f64("floatfield".to_string(), 10.0..11.0)),
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_f64(float_field, 10.0)),
+                Bound::Excluded(Term::from_field_f64(float_field, 11.0)),
+            )),
             9
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_f64_bounds(
-                "floatfield".to_string(),
-                Bound::Included(10.0),
-                Bound::Included(11.0)
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_f64(float_field, 10.0)),
+                Bound::Included(Term::from_field_f64(float_field, 11.0)),
             )),
             18
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_f64_bounds(
-                "floatfield".to_string(),
-                Bound::Excluded(9.0),
-                Bound::Included(10.0)
+            count_multiples(RangeQuery::new(
+                Bound::Excluded(Term::from_field_f64(float_field, 9.0)),
+                Bound::Included(Term::from_field_f64(float_field, 10.0)),
             )),
             9
         );
         assert_eq!(
-            count_multiples(RangeQuery::new_f64_bounds(
-                "floatfield".to_string(),
-                Bound::Included(9.0),
+            count_multiples(RangeQuery::new(
+                Bound::Included(Term::from_field_f64(float_field, 9.0)),
                 Bound::Unbounded
             )),
             91
