+++ src/query/range_query/range_query_u64_fastfield.rs
@@ -2,54 +2,34 @@
 //! We use this variant only if the fastfield exists, otherwise the default in `range_query` is
 //! used, which uses the term dictionary + postings.
 
+use std::net::Ipv6Addr;
 use std::ops::{Bound, RangeInclusive};
 
-use columnar::{ColumnType, HasAssociatedColumnType, MonotonicallyMappableToU64};
+use columnar::{Column, MonotonicallyMappableToU128, MonotonicallyMappableToU64};
+use common::BinarySerializable;
 
-use super::fast_field_range_query::RangeDocSet;
-use super::map_bound;
-use crate::query::{ConstScorer, EmptyScorer, Explanation, Query, Scorer, Weight};
-use crate::{DocId, DocSet, Score, SegmentReader, TantivyError};
+use super::fast_field_range_doc_set::RangeDocSet;
+use super::{map_bound, map_bound_res};
+use crate::query::range_query::range_query::inner_bound;
+use crate::query::{AllScorer, ConstScorer, EmptyScorer, Explanation, Query, Scorer, Weight};
+use crate::schema::{Field, Type};
+use crate::{DocId, DocSet, Score, SegmentReader, TantivyError, Term};
 
 /// `FastFieldRangeWeight` uses the fast field to execute range queries.
 #[derive(Clone, Debug)]
 pub struct FastFieldRangeWeight {
-    field: String,
-    lower_bound: Bound<u64>,
-    upper_bound: Bound<u64>,
-    column_type_opt: Option<ColumnType>,
+    lower_bound: Bound<Term>,
+    upper_bound: Bound<Term>,
+    field: Field,
 }
 
 impl FastFieldRangeWeight {
-    /// Create a new FastFieldRangeWeight, using the u64 representation of any fast field.
-    pub(crate) fn new_u64_lenient(
-        field: String,
-        lower_bound: Bound<u64>,
-        upper_bound: Bound<u64>,
-    ) -> Self {
-        let lower_bound = map_bound(&lower_bound, |val| *val);
-        let upper_bound = map_bound(&upper_bound, |val| *val);
+    /// Create a new FastFieldRangeWeight
+    pub(crate) fn new(field: Field, lower_bound: Bound<Term>, upper_bound: Bound<Term>) -> Self {
         Self {
-            field,
             lower_bound,
             upper_bound,
-            column_type_opt: None,
-        }
-    }
-
-    /// Create a new `FastFieldRangeWeight` for a range of a u64-mappable type .
-    pub fn new<T: HasAssociatedColumnType + MonotonicallyMappableToU64>(
-        field: String,
-        lower_bound: Bound<T>,
-        upper_bound: Bound<T>,
-    ) -> Self {
-        let lower_bound = map_bound(&lower_bound, |val| val.to_u64());
-        let upper_bound = map_bound(&upper_bound, |val| val.to_u64());
-        Self {
             field,
-            lower_bound,
-            upper_bound,
-            column_type_opt: Some(T::column_type()),
         }
     }
 }
@@ -65,30 +45,86 @@ impl Query for FastFieldRangeWeight {
 
 impl Weight for FastFieldRangeWeight {
     fn scorer(&self, reader: &SegmentReader, boost: Score) -> crate::Result<Box<dyn Scorer>> {
-        let fast_field_reader = reader.fast_fields();
-        let column_type_opt: Option<[ColumnType; 1]> =
-            self.column_type_opt.map(|column_type| [column_type]);
-        let column_type_opt_ref: Option<&[ColumnType]> = column_type_opt
-            .as_ref()
-            .map(|column_types| column_types.as_slice());
-        let Some((column, _)) =
-            fast_field_reader.u64_lenient_for_type(column_type_opt_ref, &self.field)?
-        else {
-            return Ok(Box::new(EmptyScorer));
-        };
-        #[allow(clippy::reversed_empty_ranges)]
-        let value_range = bound_to_value_range(
-            &self.lower_bound,
-            &self.upper_bound,
-            column.min_value(),
-            column.max_value(),
-        )
-        .unwrap_or(1..=0); // empty range
-        if value_range.is_empty() {
-            return Ok(Box::new(EmptyScorer));
+        // Check if both bounds are Bound::Unbounded
+        if self.lower_bound == Bound::Unbounded && self.upper_bound == Bound::Unbounded {
+            return Ok(Box::new(AllScorer::new(reader.max_doc())));
         }
-        let docset = RangeDocSet::new(value_range, column);
-        Ok(Box::new(ConstScorer::new(docset, boost)))
+        let field_name = reader.schema().get_field_name(self.field);
+        let field_type = reader.schema().get_field_entry(self.field).field_type();
+        if field_type.is_ip_addr() {
+            let parse_ip_from_bytes = |term: &Term| {
+                let ip_u128_bytes: [u8; 16] =
+                    term.serialized_value_bytes().try_into().map_err(|_| {
+                        crate::TantivyError::InvalidArgument(
+                            "Expected 8 bytes for ip address".to_string(),
+                        )
+                    })?;
+                let ip_u128 = u128::from_be_bytes(ip_u128_bytes);
+                crate::Result::<Ipv6Addr>::Ok(Ipv6Addr::from_u128(ip_u128))
+            };
+            let lower_bound = map_bound_res(&self.lower_bound, parse_ip_from_bytes)?;
+            let upper_bound = map_bound_res(&self.upper_bound, parse_ip_from_bytes)?;
+
+            let Some(ip_addr_column): Option<Column<Ipv6Addr>> =
+                reader.fast_fields().column_opt(field_name)?
+            else {
+                return Ok(Box::new(EmptyScorer));
+            };
+            let value_range = bound_to_value_range_ip(
+                &lower_bound,
+                &upper_bound,
+                ip_addr_column.min_value(),
+                ip_addr_column.max_value(),
+            );
+            let docset = RangeDocSet::new(value_range, ip_addr_column);
+            Ok(Box::new(ConstScorer::new(docset, boost)))
+        } else {
+            assert!(
+                maps_to_u64_fastfield(field_type.value_type()),
+                "{:?}",
+                field_type
+            );
+
+            let term = inner_bound(&self.lower_bound)
+                .or(inner_bound(&self.upper_bound))
+                .expect("At least one bound must be set");
+            assert_eq!(
+                term.typ(),
+                field_type.value_type(),
+                "Field is of type {:?}, but got term of type {:?}",
+                field_type,
+                term.typ()
+            );
+
+            let parse_from_bytes = |term: &Term| {
+                u64::from_be(
+                    BinarySerializable::deserialize(&mut &term.serialized_value_bytes()[..])
+                        .unwrap(),
+                )
+            };
+
+            let lower_bound = map_bound(&self.lower_bound, parse_from_bytes);
+            let upper_bound = map_bound(&self.upper_bound, parse_from_bytes);
+
+            let fast_field_reader = reader.fast_fields();
+            let Some((column, _)) = fast_field_reader.u64_lenient_for_type(None, field_name)?
+            else {
+                return Ok(Box::new(EmptyScorer));
+            };
+            #[allow(clippy::reversed_empty_ranges)]
+            let value_range = bound_to_value_range(
+                &lower_bound,
+                &upper_bound,
+                column.min_value(),
+                column.max_value(),
+            )
+            .unwrap_or(1..=0); // empty range
+            if value_range.is_empty() {
+                return Ok(Box::new(EmptyScorer));
+            }
+            let docset = RangeDocSet::new(value_range, column);
+            Ok(Box::new(ConstScorer::new(docset, boost)))
+        }
     }
 
     fn explain(&self, reader: &SegmentReader, doc: DocId) -> crate::Result<Explanation> {
@@ -102,6 +138,35 @@ impl Weight for FastFieldRangeWeight {
 
         Ok(explanation)
     }
+}
+
+/// Returns true if the type maps to a u64 fast field
+pub(crate) fn maps_to_u64_fastfield(typ: Type) -> bool {
+    match typ {
+        Type::U64 | Type::I64 | Type::F64 | Type::Bool | Type::Date => true,
+        Type::IpAddr => false,
+        Type::Str | Type::Facet | Type::Bytes | Type::Json => false,
+    }
+}
+
+fn bound_to_value_range_ip(
+    lower_bound: &Bound<Ipv6Addr>,
+    upper_bound: &Bound<Ipv6Addr>,
+    min_value: Ipv6Addr,
+    max_value: Ipv6Addr,
+) -> RangeInclusive<Ipv6Addr> {
+    let start_value = match lower_bound {
+        Bound::Included(ip_addr) => *ip_addr,
+        Bound::Excluded(ip_addr) => Ipv6Addr::from(ip_addr.to_u128() + 1),
+        Bound::Unbounded => min_value,
+    };
+
+    let end_value = match upper_bound {
+        Bound::Included(ip_addr) => *ip_addr,
+        Bound::Excluded(ip_addr) => Ipv6Addr::from(ip_addr.to_u128() - 1),
+        Bound::Unbounded => max_value,
+    };
+    start_value..=end_value
 }
 
 // Returns None, if the range cannot be converted to a inclusive range (which equals to a empty
@@ -141,7 +206,7 @@ pub mod tests {
     use crate::query::range_query::range_query_u64_fastfield::FastFieldRangeWeight;
     use crate::query::{QueryParser, Weight};
     use crate::schema::{NumericOptions, Schema, SchemaBuilder, FAST, INDEXED, STORED, STRING};
-    use crate::{Index, IndexWriter, TERMINATED};
+    use crate::{Index, IndexWriter, Term, TERMINATED};
 
     #[derive(Clone, Debug)]
     pub struct Doc {
@@ -213,10 +278,10 @@ pub mod tests {
         writer.add_document(doc!(field=>52_000u64)).unwrap();
         writer.commit().unwrap();
         let searcher = index.reader().unwrap().searcher();
-        let range_query = FastFieldRangeWeight::new_u64_lenient(
-            "test_field".to_string(),
-            Bound::Included(50_000),
-            Bound::Included(50_002),
+        let range_query = FastFieldRangeWeight::new(
+            field,
+            Bound::Included(Term::from_field_u64(field, 50_000)),
+            Bound::Included(Term::from_field_u64(field, 50_002)),
         );
         let scorer = range_query
             .scorer(searcher.segment_reader(0), 1.0f32)
@@ -388,6 +453,202 @@ pub mod tests {
         }
         if samples.len() > 2 {
             test_sample(vec![samples[1].clone(), samples[2].clone()]);
+        }
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+pub mod ip_range_tests {
+    use proptest::prelude::ProptestConfig;
+    use proptest::strategy::Strategy;
+    use proptest::{prop_oneof, proptest};
+
+    use super::*;
+    use crate::collector::Count;
+    use crate::query::QueryParser;
+    use crate::schema::{Schema, FAST, INDEXED, STORED, STRING};
+    use crate::{Index, IndexWriter};
+
+    #[derive(Clone, Debug)]
+    pub struct Doc {
+        pub id: String,
+        pub ip: Ipv6Addr,
+    }
+
+    fn operation_strategy() -> impl Strategy<Value = Doc> {
+        prop_oneof![
+            (0u64..10_000u64).prop_map(doc_from_id_1),
+            (1u64..10_000u64).prop_map(doc_from_id_2),
+        ]
+    }
+
+    pub fn doc_from_id_1(id: u64) -> Doc {
+        let id = id * 1000;
+        Doc {
+            // ip != id
+            id: id.to_string(),
+            ip: Ipv6Addr::from_u128(id as u128),
+        }
+    }
+    fn doc_from_id_2(id: u64) -> Doc {
+        let id = id * 1000;
+        Doc {
+            // ip != id
+            id: (id - 1).to_string(),
+            ip: Ipv6Addr::from_u128(id as u128),
+        }
+    }
+
+    proptest! {
+        #![proptest_config(ProptestConfig::with_cases(10))]
+        #[test]
+        fn test_ip_range_for_docs_prop(ops in proptest::collection::vec(operation_strategy(), 1..1000)) {
+            assert!(test_ip_range_for_docs(&ops).is_ok());
+        }
+    }
+
+    #[test]
+    fn test_ip_range_regression1() {
+        let ops = &[doc_from_id_1(0)];
+        assert!(test_ip_range_for_docs(ops).is_ok());
+    }
+
+    #[test]
+    fn test_ip_range_regression2() {
+        let ops = &[
+            doc_from_id_1(52),
+            doc_from_id_1(63),
+            doc_from_id_1(12),
+            doc_from_id_2(91),
+            doc_from_id_2(33),
+        ];
+        assert!(test_ip_range_for_docs(ops).is_ok());
+    }
+
+    #[test]
+    fn test_ip_range_regression3() {
+        let ops = &[doc_from_id_1(1), doc_from_id_1(2), doc_from_id_1(3)];
+        assert!(test_ip_range_for_docs(ops).is_ok());
+    }
+
+    #[test]
+    fn test_ip_range_regression3_simple() {
+        let mut schema_builder = Schema::builder();
+        let ips_field = schema_builder.add_ip_addr_field("ips", FAST | INDEXED);
+        let schema = schema_builder.build();
+        let index = Index::create_in_ram(schema);
+        let mut writer: IndexWriter = index.writer_for_tests().unwrap();
+        let ip_addrs: Vec<Ipv6Addr> = [1000, 2000, 3000]
+            .into_iter()
+            .map(Ipv6Addr::from_u128)
+            .collect();
+        for &ip_addr in &ip_addrs {
+            writer
+                .add_document(doc!(ips_field=>ip_addr, ips_field=>ip_addr))
+                .unwrap();
+        }
+        writer.commit().unwrap();
+        let searcher = index.reader().unwrap().searcher();
+        let range_weight = FastFieldRangeWeight::new(
+            ips_field,
+            Bound::Included(Term::from_field_ip_addr(ips_field, ip_addrs[1])),
+            Bound::Included(Term::from_field_ip_addr(ips_field, ip_addrs[2])),
+        );
+
+        let count =
+            crate::query::weight::Weight::count(&range_weight, searcher.segment_reader(0)).unwrap();
+        assert_eq!(count, 2);
+    }
+
+    pub fn create_index_from_ip_docs(docs: &[Doc]) -> Index {
+        let mut schema_builder = Schema::builder();
+        let ip_field = schema_builder.add_ip_addr_field("ip", STORED | FAST);
+        let ips_field = schema_builder.add_ip_addr_field("ips", FAST | INDEXED);
+        let text_field = schema_builder.add_text_field("id", STRING | STORED);
+        let schema = schema_builder.build();
+        let index = Index::create_in_ram(schema);
+
+        {
+            let mut index_writer = index.writer_with_num_threads(2, 60_000_000).unwrap();
+            for doc in docs.iter() {
+                index_writer
+                    .add_document(doc!(
+                        ips_field => doc.ip,
+                        ips_field => doc.ip,
+                        ip_field => doc.ip,
+                        text_field => doc.id.to_string(),
+                    ))
+                    .unwrap();
+            }
+
+            index_writer.commit().unwrap();
+        }
+        index
+    }
+
+    fn test_ip_range_for_docs(docs: &[Doc]) -> crate::Result<()> {
+        let index = create_index_from_ip_docs(docs);
+        let reader = index.reader().unwrap();
+        let searcher = reader.searcher();
+
+        let get_num_hits = |query| searcher.search(&query, &Count).unwrap();
+        let query_from_text = |text: &str| {
+            QueryParser::for_index(&index, vec![])
+                .parse_query(text)
+                .unwrap()
+        };
+
+        let gen_query_inclusive = |field: &str, ip_range: &RangeInclusive<Ipv6Addr>| {
+            format!("{field}:[{} TO {}]", ip_range.start(), ip_range.end())
+        };
+
+        let test_sample = |sample_docs: &[Doc]| {
+            let mut ips: Vec<Ipv6Addr> = sample_docs.iter().map(|doc| doc.ip).collect();
+            ips.sort();
+            let ip_range = ips[0]..=ips[1];
+            let expected_num_hits = docs
+                .iter()
+                .filter(|doc| (ips[0]..=ips[1]).contains(&doc.ip))
+                .count();
+
+            let query = gen_query_inclusive("ip", &ip_range);
+            assert_eq!(get_num_hits(query_from_text(&query)), expected_num_hits);
+
+            let query = gen_query_inclusive("ips", &ip_range);
+            assert_eq!(get_num_hits(query_from_text(&query)), expected_num_hits);
+
+            // Intersection search
+            let id_filter = sample_docs[0].id.to_string();
+            let expected_num_hits = docs
+                .iter()
+                .filter(|doc| ip_range.contains(&doc.ip) && doc.id == id_filter)
+                .count();
+            let query = format!(
+                "{} AND id:{}",
+                gen_query_inclusive("ip", &ip_range),
+                &id_filter
+            );
+            assert_eq!(get_num_hits(query_from_text(&query)), expected_num_hits);
+
+            // Intersection search on multivalue ip field
+            let id_filter = sample_docs[0].id.to_string();
+            let query = format!(
+                "{} AND id:{}",
+                gen_query_inclusive("ips", &ip_range),
+                &id_filter
+            );
+            assert_eq!(get_num_hits(query_from_text(&query)), expected_num_hits);
+        };
+
+        test_sample(&[docs[0].clone(), docs[0].clone()]);
+        if docs.len() > 1 {
+            test_sample(&[docs[0].clone(), docs[1].clone()]);
+            test_sample(&[docs[1].clone(), docs[1].clone()]);
+        }
+        if docs.len() > 2 {
+            test_sample(&[docs[1].clone(), docs[2].clone()]);
         }
 
         Ok(())
@@ -599,5 +860,244 @@ mod bench {
     fn bench_id_range_hit_90_percent_intersect_with_1_percent_multi(bench: &mut Bencher) {
         let index = get_index_0_to_100();
         bench.iter(|| execute_query("ids", get_90_percent(), "AND id_name:veryfew", &index));
+    }
+}
+
+#[cfg(all(test, feature = "unstable"))]
+mod bench_ip {
+
+    use rand::rngs::StdRng;
+    use rand::{Rng, SeedableRng};
+    use test::Bencher;
+
+    use super::ip_range_tests::*;
+    use super::*;
+    use crate::collector::Count;
+    use crate::query::QueryParser;
+    use crate::Index;
+
+    fn get_index_0_to_100() -> Index {
+        let mut rng = StdRng::from_seed([1u8; 32]);
+        let num_vals = 100_000;
+        let docs: Vec<_> = (0..num_vals)
+            .map(|_i| {
+                let id = if rng.gen_bool(0.01) {
+                    "veryfew".to_string() // 1%
+                } else if rng.gen_bool(0.1) {
+                    "few".to_string() // 9%
+                } else {
+                    "many".to_string() // 90%
+                };
+                Doc {
+                    id,
+                    // Multiply by 1000, so that we create many buckets in the compact space
+                    // The benches depend on this range to select n-percent of elements with the
+                    // methods below.
+                    ip: Ipv6Addr::from_u128(rng.gen_range(0..100) * 1000),
+                }
+            })
+            .collect();
+
+        create_index_from_ip_docs(&docs)
+    }
+
+    fn get_90_percent() -> RangeInclusive<Ipv6Addr> {
+        let start = Ipv6Addr::from_u128(0);
+        let end = Ipv6Addr::from_u128(90 * 1000);
+        start..=end
+    }
+
+    fn get_10_percent() -> RangeInclusive<Ipv6Addr> {
+        let start = Ipv6Addr::from_u128(0);
+        let end = Ipv6Addr::from_u128(10 * 1000);
+        start..=end
+    }
+
+    fn get_1_percent() -> RangeInclusive<Ipv6Addr> {
+        let start = Ipv6Addr::from_u128(10 * 1000);
+        let end = Ipv6Addr::from_u128(10 * 1000);
+        start..=end
+    }
+
+    fn excute_query(
+        field: &str,
+        ip_range: RangeInclusive<Ipv6Addr>,
+        suffix: &str,
+        index: &Index,
+    ) -> usize {
+        let gen_query_inclusive = |from: &Ipv6Addr, to: &Ipv6Addr| {
+            format!(
+                "{}:[{} TO {}] {}",
+                field,
+                &from.to_string(),
+                &to.to_string(),
+                suffix
+            )
+        };
+
+        let query = gen_query_inclusive(ip_range.start(), ip_range.end());
+        let query_from_text = |text: &str| {
+            QueryParser::for_index(index, vec![])
+                .parse_query(text)
+                .unwrap()
+        };
+        let query = query_from_text(&query);
+        let reader = index.reader().unwrap();
+        let searcher = reader.searcher();
+        searcher.search(&query, &(Count)).unwrap()
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_90_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_10_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_1_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent_intersect_with_10_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_10_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_10_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_1_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_90_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_1_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_1_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_1_percent(), "AND id:veryfew", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent_intersect_with_90_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_10_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_90_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_90_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_10_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_90_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_1_percent(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ip", get_90_percent(), "AND id:veryfew", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_90_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_10_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_1_percent(), "", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent_intersect_with_10_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_10_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_10_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_1_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_90_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_1_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_1_percent_intersect_with_1_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_1_percent(), "AND id:veryfew", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_10_percent_intersect_with_90_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_10_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_90_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_90_percent(), "AND id:many", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_10_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_90_percent(), "AND id:few", &index));
+    }
+
+    #[bench]
+    fn bench_ip_range_hit_90_percent_intersect_with_1_percent_multi(bench: &mut Bencher) {
+        let index = get_index_0_to_100();
+
+        bench.iter(|| excute_query("ips", get_90_percent(), "AND id:veryfew", &index));
     }
 }
